requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.8.3
google-genai>=1.46.0
python-multipart==0.0.6
pydantic>=2.7.0,<3.0.0
asyncio==3.4.3
websockets>=13.0.0,<15.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx>=0.24.0
beautifulsoup4==4.12.3
requests==2.31.0